[PBA-2878] Rollback scrubber handleTouchMove()
The PR #253 introduced an additional line of code in the handleTouchMove function
for the bottomOverlay component. I'm removing it as it is causing the bug described
in PBA-2878.

We may put it back again but considering that it breaks the UX in iOS when you try
to drag the scrub bar for the volume component, as it also drags the progress bar.