[PBA-5493] Deal with boolean values only for showUpNext setting
Being cautious about how we deal with the settings provided in skin.json, showUpNext should be a bool value but the customer can provide anything they want. Right now if they provide something different than a boolean strange behavior may occur with showing the upNext widget or not showing it, and also showing the discovery on end screen. If customers use the skin.json as it should be they won’t encounter any problems.